resolve functions
