+++ src/lib/Functions/pasteData.ts
@@ -1,9 +1,53 @@
-import { Cell, Compatible } from '../Model/PublicModel';
-import { State } from '../Model/State';
-import { Range } from '../Model/Range';
-import { tryAppendChangeHavingGroupId } from './tryAppendChangeHavingGroupId';
+import { Cell, Compatible } from "../Model/PublicModel";
+import { State } from "../Model/State";
+import { Location } from "../Model/InternalModel";
+import { tryAppendChangeHavingGroupId } from "./tryAppendChangeHavingGroupId";
+import { getActiveSelectedRange } from "./getActiveSelectedRange";
+import { newLocation } from "./newLocation";
 
-export function pasteData(state: State, activeSelectedRange: Range, cellToPaste: Compatible<Cell>): State {
-    const location = { row: activeSelectedRange.first.row, column: activeSelectedRange.first.column };
-    return tryAppendChangeHavingGroupId(state, location, cellToPaste);
+export function pasteData(state: State, rows: Compatible<Cell>[][]): State {
+  const activeSelectedRange = getActiveSelectedRange(state);
+  if (rows.length === 1 && rows[0].length === 1) {
+    // pasting single cell on active selected range
+    activeSelectedRange.rows.forEach((row) =>
+      activeSelectedRange.columns.forEach((column) => {
+        state = tryAppendChangeHavingGroupId(
+          state, 
+          newLocation(row, column),
+          rows[0][0]
+        ) as State;
+      })
+    );
+  } else {
+    let lastLocation: Location | undefined;
+    const cellMatrix = state.cellMatrix;
+    rows.forEach((row, ri) =>
+      row.forEach((cell, ci) => {
+        const rowIdx = activeSelectedRange.first.row.idx + ri;
+        const columnIdx = activeSelectedRange.first.column.idx + ci;
+        if (
+          rowIdx <= cellMatrix.last.row.idx &&
+          columnIdx <= cellMatrix.last.column.idx
+        ) {
+          lastLocation = cellMatrix.getLocation(rowIdx, columnIdx);
+          state = tryAppendChangeHavingGroupId(
+            state,
+            lastLocation,
+            cell
+          ) as State;
+        }
+      })
+    );
+    if (!lastLocation) {
+      return state;
+    }
+    return {
+      ...state,
+      selectedRanges: [
+        cellMatrix.getRange(activeSelectedRange.first, lastLocation),
+      ],
+      activeSelectedRangeIdx: 0,
+    };
+  }
+  return state;
 }
